feat(mc_map): add mc_map_for_each and fix some bugs
Fix a segmentation fault in mc_hash_table_init caused by returning early
when the capacity is 0, which leads to uninitialized memory layout of
entries.

Fix the segmentation fault that may occur when the hash table is empty
due to mc_map_get performing a lookup without checking the length.