feat(mc_test): add mc_try_register_empty_suite
In some cases, the registration function of a test may be executed
before the registration function of the corresponding test suite.
Previously, we ensured that the test suite's registration function
executed first by setting the __attribute__((constructor)) parameter to
101. Now, we add the mc_try_register_empty_suite function to handle this
case. If the test suite does not exist, it will be created and
registered.